+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/payments/refunds/IssueRefundViewModel.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.refunds
+package com.woocommerce.android.ui.payments.refunds
 
 import android.os.Parcelable
 import androidx.lifecycle.LiveData
@@ -34,21 +34,21 @@ import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.orders.details.OrderDetailRepository
 import com.woocommerce.android.ui.payments.cardreader.InPersonPaymentsCanadaFeatureFlag
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.InputValidationState.TOO_HIGH
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.InputValidationState.TOO_LOW
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.InputValidationState.VALID
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.HideValidationError
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.OpenUrl
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.ShowNumberPicker
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundAmountDialog
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundConfirmation
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundSummary
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.IssueRefundEvent.ShowValidationError
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.RefundType.AMOUNT
-import com.woocommerce.android.ui.refunds.IssueRefundViewModel.RefundType.ITEMS
-import com.woocommerce.android.ui.refunds.RefundFeeListAdapter.FeeRefundListItem
-import com.woocommerce.android.ui.refunds.RefundProductListAdapter.ProductRefundListItem
-import com.woocommerce.android.ui.refunds.RefundShippingListAdapter.ShippingRefundListItem
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.InputValidationState.TOO_HIGH
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.InputValidationState.TOO_LOW
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.InputValidationState.VALID
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.HideValidationError
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.OpenUrl
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowNumberPicker
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundAmountDialog
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundConfirmation
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundSummary
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowValidationError
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.RefundType.AMOUNT
+import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.RefundType.ITEMS
+import com.woocommerce.android.ui.payments.refunds.RefundFeeListAdapter.FeeRefundListItem
+import com.woocommerce.android.ui.payments.refunds.RefundProductListAdapter.ProductRefundListItem
+import com.woocommerce.android.ui.payments.refunds.RefundShippingListAdapter.ShippingRefundListItem
 import com.woocommerce.android.util.CoroutineDispatchers
 import com.woocommerce.android.util.CurrencyFormatter
 import com.woocommerce.android.util.max
