+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/payments/refunds/RefundsArgs.kt
@@ -1,0 +1,3 @@
+package com.woocommerce.android.ui.payments.refunds
+
+typealias RefundsArgs = IssueRefundFragmentArgs
