--- WooCommerce/src/main/kotlin/com/woocommerce/android/ui/refunds/RefundsArgs.kt
@@ -1,3 +1,0 @@
-package com.woocommerce.android.ui.refunds
-
-typealias RefundsArgs = IssueRefundFragmentArgs
