+++ WooCommerce/src/main/res/navigation/nav_graph_orders.xml
@@ -143,7 +143,7 @@
     </dialog>
     <fragment
         android:id="@+id/refundDetailFragment"
-        android:name="com.woocommerce.android.ui.refunds.RefundDetailFragment"
+        android:name="com.woocommerce.android.ui.payments.refunds.RefundDetailFragment"
         android:label="RefundDetailFragment">
         <argument
             android:name="orderId"
